Rename repository to service to follow the blazor naming convention
