+++ src/Wasserwacht.DigitalGuardBook.Common.Ui/StationInfo.razor
@@ -1,5 +1,5 @@
-﻿@using Wasserwacht.DigitalGuardBook.Common.Logic.Repositories
-@inject StationRepository StationService
+﻿@using Wasserwacht.DigitalGuardBook.Common.Logic.Services
+@inject StationService StationService
 
 <div class="card">
     <div class="card-body">
