+++ src/Wasserwacht.DigitalGuardBook.Ui/Shared/Header.razor
@@ -1,5 +1,5 @@
-﻿@using Wasserwacht.DigitalGuardBook.Common.Logic.Repositories
-@inject StationRepository StationService
+﻿@using Wasserwacht.DigitalGuardBook.Common.Logic.Services
+@inject StationService StationService
 
 <nav class="navbar navbar-expand-md sticky-top navbar-light bg-light">
     <div class="container-fluid">
