+++ src/Wasserwacht.DigitalGuardBook.Ui/Startup.cs
@@ -35,7 +35,7 @@ namespace Wasserwacht.DigitalGuardBook.Ui
             services.AddServerSideBlazor();
             services.AddScoped<AuthenticationStateProvider, RevalidatingIdentityAuthenticationStateProvider<IdentityUser>>();
             services.AddDatabaseDeveloperPageExceptionFilter();
-            services.AddSingleton<Common.Logic.Repositories.StationRepository>();
+            services.AddSingleton<Common.Logic.Services.StationService>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
